fix: support state sync on Windows
