+++ internal/sync/state/state.go
@@ -282,22 +282,6 @@ func containsParentTraversal(components []string) bool {
 	return false
 }
 
-func syncDirectory(path string) error {
-	dir, err := os.Open(path)
-	if err != nil {
-		return fmt.Errorf("failed to open directory: %w", err)
-	}
-	if err := dir.Sync(); err != nil {
-		_ = dir.Close()
-		return fmt.Errorf("failed to sync directory: %w", err)
-	}
-	if err := dir.Close(); err != nil {
-		return fmt.Errorf("failed to close directory: %w", err)
-	}
-
-	return nil
-}
-
 func (s *State) UpdateBook(bookID string, progress float64, status string) bool {
 	s.mu.Lock()
 	defer s.mu.Unlock()
